refactor into more useable code
