+++ music_assistant/providers/airplay/helpers.py
@@ -4,15 +4,19 @@ from __future__ import annotations
 
 import os
 import platform
+import time
 from typing import TYPE_CHECKING
 
 from zeroconf import IPVersion
 
 from music_assistant.helpers.process import check_output
-from music_assistant.providers.airplay.constants import BROKEN_RAOP_MODELS
+from music_assistant.providers.airplay.constants import BROKEN_RAOP_MODELS, StreamingProtocol
 
 if TYPE_CHECKING:
     from zeroconf.asyncio import AsyncServiceInfo
+
+# NTP epoch delta: difference between Unix epoch (1970) and NTP epoch (1900)
+NTP_EPOCH_DELTA = 0x83AA7E80  # 2208988800 seconds
 
 
 def convert_airplay_volume(value: float) -> int:
@@ -111,12 +115,22 @@ def is_broken_raop_model(manufacturer: str, model: str) -> bool:
     return False
 
 
-async def get_cli_binary(version: int) -> str:
-    """Find the correct raop/airplay binary belonging to the platform."""
+async def get_cli_binary(protocol: StreamingProtocol) -> str:
+    """Find the correct raop/airplay binary belonging to the platform.
+
+    Args:
+        protocol: The streaming protocol (RAOP or AIRPLAY2)
+
+    Returns:
+        Path to the CLI binary
+
+    Raises:
+        RuntimeError: If the binary cannot be found
+    """
 
     async def check_binary(cli_path: str) -> str | None:
         try:
-            if version == 1:
+            if protocol == StreamingProtocol.RAOP:
                 args = [
                     cli_path,
                     "-check",
@@ -132,9 +146,11 @@ async def get_cli_binary(version: int) -> str:
                 ]
 
             returncode, output = await check_output(*args)
-            if (version == 1 and returncode == 0 and output.strip().decode() == passing_output) or (
-                version == 2 and returncode == 0
-            ):
+            if (
+                protocol == StreamingProtocol.RAOP
+                and returncode == 0
+                and output.strip().decode() == passing_output
+            ) or (protocol == StreamingProtocol.AIRPLAY2 and returncode == 0):
                 return cli_path
         except OSError:
             pass
@@ -143,17 +159,82 @@ async def get_cli_binary(version: int) -> str:
     base_path = os.path.join(os.path.dirname(__file__), "bin")
     system = platform.system().lower().replace("darwin", "macos")
     architecture = platform.machine().lower()
-    if version == 1:
+
+    if protocol == StreamingProtocol.RAOP:
         package = "cliraop"
-    elif version == 2:
+    elif protocol == StreamingProtocol.AIRPLAY2:
         package = "cliap2"
     else:
-        raise RuntimeError(f"Unsupported CLI binary version requested: {version}")
+        raise RuntimeError(f"Unsupported streaming protocol requested: {protocol}")
 
     if bridge_binary := await check_binary(
         os.path.join(base_path, f"{package}-{system}-{architecture}")
     ):
         return bridge_binary
 
-    msg = f"Unable to locate RAOP/AirPlay2 Play binary {package} for {system}/{architecture}"
+    msg = (
+        f"Unable to locate {protocol.name} CLI stream binary {package} for {system}/{architecture}"
+    )
     raise RuntimeError(msg)
+
+
+def get_ntp_timestamp() -> int:
+    """
+    Get current NTP timestamp (64-bit).
+
+    Returns:
+        int: 64-bit NTP timestamp (upper 32 bits = seconds, lower 32 bits = fraction)
+    """
+    # Get current Unix timestamp with microsecond precision
+    current_time = time.time()
+
+    # Split into seconds and microseconds
+    seconds = int(current_time)
+    microseconds = int((current_time - seconds) * 1_000_000)
+
+    # Convert to NTP epoch (add offset from 1970 to 1900)
+    ntp_seconds = seconds + NTP_EPOCH_DELTA
+
+    # Convert microseconds to NTP fraction (2^32 parts per second)
+    # fraction = (microseconds * 2^32) / 1_000_000
+    ntp_fraction = int((microseconds << 32) / 1_000_000)
+
+    # Combine into 64-bit value
+    return (ntp_seconds << 32) | ntp_fraction
+
+
+def ntp_to_seconds_fraction(ntp_timestamp: int) -> tuple[int, int]:
+    """
+    Split NTP timestamp into seconds and fraction components.
+
+    Args:
+        ntp_timestamp: 64-bit NTP timestamp
+
+    Returns:
+        tuple: (seconds, fraction)
+    """
+    seconds = ntp_timestamp >> 32
+    fraction = ntp_timestamp & 0xFFFFFFFF
+    return seconds, fraction
+
+
+def ntp_to_unix_time(ntp_timestamp: int) -> float:
+    """
+    Convert NTP timestamp to Unix timestamp (float).
+
+    Args:
+        ntp_timestamp: 64-bit NTP timestamp
+
+    Returns:
+        float: Unix timestamp (seconds since 1970-01-01)
+    """
+    seconds = ntp_timestamp >> 32
+    fraction = ntp_timestamp & 0xFFFFFFFF
+
+    # Convert back to Unix epoch
+    unix_seconds = seconds - NTP_EPOCH_DELTA
+
+    # Convert fraction to microseconds
+    microseconds = (fraction * 1_000_000) >> 32
+
+    return unix_seconds + (microseconds / 1_000_000)
