+++ music_assistant/providers/airplay/protocols/__init__.py
@@ -1,0 +1,1 @@
+"""AirPlay (streaming) Protocols."""
